Extracted BuildServer implementation more properly
